Fix cross platform details in json_printf
1. support native size_t and (u)int64_t formatters
2. fix %lld on 32-bit unices
3. fix x86 32-bit (e.g. linux and 32-bit osx) var_arg portability

The trick used to move to the next argument on windows was wrong:
a) it was marked as a windows-only hack but instead it would affect any
x86-32 ABI because va_arg there is a simple pointer and thus passed by
value. And so it would likely be on any platform where arguments are
solely passed on the stack. Anyway, it's compiler and ABI dependent.
b) it assumed that any type was wide as `void*` which obviously is not
true, e.g.: in `json_printf(out, "%lld %d", a, b)` if we consume a void*
when processing `a` we won't land on `b`.

The solution is to actually parse the format.
We're more likely to get all the printf types right than to get all
the embedded platform vararg right. Or, to put it differently: we care
less if we get obscure formats wrong, but we care if on a new port the
basic %c, %d, %zu, %lu, %llu formats suddenly don't work.

This could turn into a complete fprintf shim. I left the part that delegates
the actual modifier parsing to system snprintf.

Added 32bit unix test

PUBLISHED_FROM=e53f30b385c63050e74f641983c2f9abcc06c336