process %s and %.*s internally in json_vprintf()
large strings (> 20 chars) need an additional allocation of memory
in order to provide a large enough buffer to be processed with
vsnprintf(). With this patch %s and %.*s are processed internally
in json_vprintf() and not delegated to vsnprintf() anymore. So in
theory json_vprintf() should work without calling malloc().
Nevertheless the delegating section to vnsprintf() stays intact
for all the other format strings. For format string / argument
combinations that produce outputs of longer than 20 chars, the
double buffering method is left in place.