Add a .gitignore file
This initial version instructs git to ignore build artifacts
from the environment which I am using. Your environment might
leave different artifacts, in which case the file might need
to be extended.

This is particularly relevant for use as a submodule, which
otherwise is always shown as dirty after a build.